# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountMode = typeof AccountMode[keyof typeof AccountMode];


export const AccountMode = {
  live: 'live',
  demo: 'demo',
  disconnected: 'disconnected',
} as const;

export interface Account {
  connected: boolean;
  /** @nullable */
  email: string | null;
  mode: AccountMode;
  /** @nullable */
  lastSyncedAt?: string | null;
}

export interface Category {
  id: string;
  name: string;
  /** @minimum 0 */
  count: number;
  color: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  share?: number;
}

export interface Dashboard {
  account: Account;
  /** @minimum 0 */
  totalEmails: number;
  /** @minimum 0 */
  unreadEmails: number;
  /** @minimum 0 */
  recurringSenders: number;
  categories: Category[];
  /** @nullable */
  lastScan: string | null;
  /** @minimum 0 */
  labelsApplied: number;
  privacyNote?: string;
}

export type ScanInputScope = typeof ScanInputScope[keyof typeof ScanInputScope];


export const ScanInputScope = {
  inbox: 'inbox',
  unread: 'unread',
  recent: 'recent',
} as const;

export interface ScanInput {
  scope?: ScanInputScope;
  /**
     * @minimum 1
     * @maximum 10000
     */
  maxMessages?: number;
}

export type ScanJobStatus = typeof ScanJobStatus[keyof typeof ScanJobStatus];


export const ScanJobStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ScanJob {
  id: string;
  status: ScanJobStatus;
  /** @minimum 0 */
  totalMessages: number;
  /** @minimum 0 */
  processedMessages: number;
  /** @minimum 0 */
  failedMessages: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  currentStep: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface Classification {
  id: string;
  gmailMessageId: string;
  gmailThreadId?: string;
  senderEmail: string;
  senderDomain: string;
  subjectPreview: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  classifiedAt: string;
  included: boolean;
}

export interface ClassificationUpdate {
  category?: string;
  included?: boolean;
}

export interface LabelApplyInput {
  /** @minItems 1 */
  classificationIds: string[];
}

export type LabelApplyResultStatus = typeof LabelApplyResultStatus[keyof typeof LabelApplyResultStatus];


export const LabelApplyResultStatus = {
  queued: 'queued',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface LabelApplyResult {
  status: LabelApplyResultStatus;
  /** @minimum 0 */
  labelsCreated: number;
  /** @minimum 0 */
  messagesLabeled: number;
  /** @minimum 0 */
  failedMessages: number;
  message?: string;
}

export type ListClassificationsParams = {
category?: string;
sender?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};