# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountMode } from './accountMode';

export interface Account {
  connected: boolean;
  /** @nullable */
  email: string | null;
  mode: AccountMode;
  /** @nullable */
  lastSyncedAt?: string | null;
}