# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  name: string;
  /** @minimum 0 */
  count: number;
  color: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  share?: number;
}