# lib/api-zod/src/generated/types/classification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */

export interface Classification {
  id: string;
  gmailMessageId: string;
  gmailThreadId?: string;
  senderEmail: string;
  senderDomain: string;
  subjectPreview: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  classifiedAt: string;
  included: boolean;
}