# lib/api-zod/src/generated/types/classificationUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassificationUpdate {
  category?: string;
  included?: boolean;
}