# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from './account';
import type { Category } from './category';

export interface Dashboard {
  account: Account;
  /** @minimum 0 */
  totalEmails: number;
  /** @minimum 0 */
  unreadEmails: number;
  /** @minimum 0 */
  recurringSenders: number;
  categories: Category[];
  /** @nullable */
  lastScan: string | null;
  /** @minimum 0 */
  labelsApplied: number;
  privacyNote?: string;
}