# lib/api-zod/src/generated/types/labelApplyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */

export interface LabelApplyInput {
  /** @minItems 1 */
  classificationIds: string[];
}