# lib/api-zod/src/generated/types/labelApplyResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
import type { LabelApplyResultStatus } from './labelApplyResultStatus';

export interface LabelApplyResult {
  status: LabelApplyResultStatus;
  /** @minimum 0 */
  labelsCreated: number;
  /** @minimum 0 */
  messagesLabeled: number;
  /** @minimum 0 */
  failedMessages: number;
  message?: string;
}