# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputScope } from './scanInputScope';

export interface ScanInput {
  scope?: ScanInputScope;
  /**
     * @minimum 1
     * @maximum 10000
     */
  maxMessages?: number;
}