# lib/api-zod/src/generated/types/scanJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Privacy-conscious Gmail organization and AI classification API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanJobStatus } from './scanJobStatus';

export interface ScanJob {
  id: string;
  status: ScanJobStatus;
  /** @minimum 0 */
  totalMessages: number;
  /** @minimum 0 */
  processedMessages: number;
  /** @minimum 0 */
  failedMessages: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  currentStep: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}